Proxy to rubygems if a gem can't be found
One of the motivations for projects to vendor all their gems was so that
we would have "local" copies of gems from the rubygems server in case it
ever went down.  This achieves the same thing, but instead of the
project keeping the copy in git (and keeping a separate copy for every
project that needs the gem), this setting would allow projects to just
point to our geminabox server without duplicating the gem for every
project that needed it.